requirements_windows.txt
requests
opencv-python
pydub
pyttsx3
jieba
langdetect
flask
Pillow
python-dotenv
pycryptodome
edge-tts